refactor(idp): extract workload projection helpers
